@@ -9,7 +9,7 @@ using Iot.Device.Bmxx80;
 using Iot.Device.Bmxx80.FilteringMode;
 using Iot.Device.Bmxx80.PowerMode;
 using Iot.Device.Common;
-using Iot.Units;
+using UnitsNet;
 
 namespace Iot.Device.Samples
 {
@@ -29,7 +29,7 @@ namespace Iot.Device.Samples
             // bus id on the raspberry pi 3
             const int busId = 1;
             // set this to the current sea level pressure in the area for correct altitude readings
-            var defaultSeaLevelPressure = Pressure.MeanSeaLevel;
+            var defaultSeaLevelPressure = WeatherHelper.MeanSeaLevel;
 
             var i2cSettings = new I2cConnectionSettings(busId, Bme280.DefaultI2cAddress);
             var i2cDevice = I2cDevice.Create(i2cSettings);
@@ -60,14 +60,14 @@ namespace Iot.Device.Samples
                     // var altValue = WeatherHelper.CalculateAltitude(preValue, defaultSeaLevelPressure, tempValue) which would be more performant.
                     i2CBmpe80.TryReadAltitude(defaultSeaLevelPressure, out var altValue);
 
-                    Console.WriteLine($"Temperature: {tempValue.Celsius:0.#}\u00B0C");
-                    Console.WriteLine($"Pressure: {preValue.Hectopascal:0.##}hPa");
+                    Console.WriteLine($"Temperature: {tempValue.DegreesCelsius:0.#}\u00B0C");
+                    Console.WriteLine($"Pressure: {preValue.Hectopascals:0.##}hPa");
                     Console.WriteLine($"Altitude: {altValue:0.##}m");
                     Console.WriteLine($"Relative humidity: {humValue:0.#}%");
 
                     // WeatherHelper supports more calculations, such as saturated vapor pressure, actual vapor pressure and absolute humidity.
-                    Console.WriteLine($"Heat index: {WeatherHelper.CalculateHeatIndex(tempValue, humValue).Celsius:0.#}\u00B0C");
-                    Console.WriteLine($"Dew point: {WeatherHelper.CalculateDewPoint(tempValue, humValue).Celsius:0.#}\u00B0C");
+                    Console.WriteLine($"Heat index: {WeatherHelper.CalculateHeatIndex(tempValue, humValue).DegreesCelsius:0.#}\u00B0C");
+                    Console.WriteLine($"Dew point: {WeatherHelper.CalculateDewPoint(tempValue, humValue).DegreesCelsius:0.#}\u00B0C");
                     Thread.Sleep(1000);
 
                     // change sampling and filter
@@ -92,14 +92,14 @@ namespace Iot.Device.Samples
                     // var altValue = WeatherHelper.CalculateAltitude(preValue, defaultSeaLevelPressure, tempValue) which would be more performant.
                     i2CBmpe80.TryReadAltitude(defaultSeaLevelPressure, out altValue);
 
-                    Console.WriteLine($"Temperature: {tempValue.Celsius:0.#}\u00B0C");
-                    Console.WriteLine($"Pressure: {preValue.Hectopascal:0.##}hPa");
+                    Console.WriteLine($"Temperature: {tempValue.DegreesCelsius:0.#}\u00B0C");
+                    Console.WriteLine($"Pressure: {preValue.Hectopascals:0.##}hPa");
                     Console.WriteLine($"Altitude: {altValue:0.##}m");
                     Console.WriteLine($"Relative humidity: {humValue:0.#}%");
 
                     // WeatherHelper supports more calculations, such as saturated vapor pressure, actual vapor pressure and absolute humidity.
-                    Console.WriteLine($"Heat index: {WeatherHelper.CalculateHeatIndex(tempValue, humValue).Celsius:0.#}\u00B0C");
-                    Console.WriteLine($"Dew point: {WeatherHelper.CalculateDewPoint(tempValue, humValue).Celsius:0.#}\u00B0C");
+                    Console.WriteLine($"Heat index: {WeatherHelper.CalculateHeatIndex(tempValue, humValue).DegreesCelsius:0.#}\u00B0C");
+                    Console.WriteLine($"Dew point: {WeatherHelper.CalculateDewPoint(tempValue, humValue).DegreesCelsius:0.#}\u00B0C");
                     Thread.Sleep(5000);
                 }
             }
