@@ -8,7 +8,7 @@ using System.Threading;
 using Iot.Device.Bmxx80;
 using Iot.Device.Bmxx80.PowerMode;
 using Iot.Device.Common;
-using Iot.Units;
+using UnitsNet;
 
 namespace Iot.Device.Samples
 {
@@ -27,7 +27,7 @@ namespace Iot.Device.Samples
             // The I2C bus ID on the Raspberry Pi 3.
             const int busId = 1;
             // set this to the current sea level pressure in the area for correct altitude readings
-            var defaultSeaLevelPressure = Pressure.MeanSeaLevel;
+            var defaultSeaLevelPressure = WeatherHelper.MeanSeaLevel;
 
             var i2cSettings = new I2cConnectionSettings(busId, Bme680.DefaultI2cAddress);
             var i2cDevice = I2cDevice.Create(i2cSettings);
@@ -56,14 +56,14 @@ namespace Iot.Device.Samples
                         var altValue = WeatherHelper.CalculateAltitude(preValue, defaultSeaLevelPressure, tempValue);
 
                         Console.WriteLine($"Gas resistance: {gasResistance:0.##}Ohm");
-                        Console.WriteLine($"Temperature: {tempValue.Celsius:0.#}\u00B0C");
-                        Console.WriteLine($"Pressure: {preValue.Hectopascal:0.##}hPa");
+                        Console.WriteLine($"Temperature: {tempValue.DegreesCelsius:0.#}\u00B0C");
+                        Console.WriteLine($"Pressure: {preValue.Hectopascals:0.##}hPa");
                         Console.WriteLine($"Altitude: {altValue:0.##}m");
                         Console.WriteLine($"Relative humidity: {humValue:0.#}%");
 
                         // WeatherHelper supports more calculations, such as saturated vapor pressure, actual vapor pressure and absolute humidity.
-                        Console.WriteLine($"Heat index: {WeatherHelper.CalculateHeatIndex(tempValue, humValue).Celsius:0.#}\u00B0C");
-                        Console.WriteLine($"Dew point: {WeatherHelper.CalculateDewPoint(tempValue, humValue).Celsius:0.#}\u00B0C");
+                        Console.WriteLine($"Heat index: {WeatherHelper.CalculateHeatIndex(tempValue, humValue).DegreesCelsius:0.#}\u00B0C");
+                        Console.WriteLine($"Dew point: {WeatherHelper.CalculateDewPoint(tempValue, humValue).DegreesCelsius:0.#}\u00B0C");
 
                         // when measuring the gas resistance on each cycle it is important to wait a certain interval
                         // because a heating plate is activated which will heat up the sensor without sleep, this can
@@ -96,14 +96,14 @@ namespace Iot.Device.Samples
                         var altValue = WeatherHelper.CalculateAltitude(preValue, defaultSeaLevelPressure, tempValue);
 
                         Console.WriteLine($"Gas resistance: {gasResistance:0.##}Ohm");
-                        Console.WriteLine($"Temperature: {tempValue.Celsius:0.#}\u00B0C");
-                        Console.WriteLine($"Pressure: {preValue.Hectopascal:0.##}hPa");
+                        Console.WriteLine($"Temperature: {tempValue.DegreesCelsius:0.#}\u00B0C");
+                        Console.WriteLine($"Pressure: {preValue.Hectopascals:0.##}hPa");
                         Console.WriteLine($"Altitude: {altValue:0.##}m");
                         Console.WriteLine($"Relative humidity: {humValue:0.#}%");
 
                         // WeatherHelper supports more calculations, such as saturated vapor pressure, actual vapor pressure and absolute humidity.
-                        Console.WriteLine($"Heat index: {WeatherHelper.CalculateHeatIndex(tempValue, humValue).Celsius:0.#}\u00B0C");
-                        Console.WriteLine($"Dew point: {WeatherHelper.CalculateDewPoint(tempValue, humValue).Celsius:0.#}\u00B0C");
+                        Console.WriteLine($"Heat index: {WeatherHelper.CalculateHeatIndex(tempValue, humValue).DegreesCelsius:0.#}\u00B0C");
+                        Console.WriteLine($"Dew point: {WeatherHelper.CalculateDewPoint(tempValue, humValue).DegreesCelsius:0.#}\u00B0C");
                         Thread.Sleep(1000);
                     }
 
