@@ -61,7 +61,7 @@ namespace Bno055sample
                 Console.WriteLine($"Gravity X: {gravity.X} Y: {gravity.Y} Z: {gravity.Z}");
                 var qua = bno055Sensor.Quaternion;
                 Console.WriteLine($"Quaternion X: {qua.X} Y: {qua.Y} Z: {qua.Z} W: {qua.W}");
-                var temp = bno055Sensor.Temperature.Celsius;
+                var temp = bno055Sensor.Temperature.DegreesCelsius;
                 Console.WriteLine($"Temperature: {temp} °C");
                 Thread.Sleep(100);
             }
