@@ -18,7 +18,7 @@ namespace Iot.Device.CpuTemperature.Samples
             {
                 if (cpuTemperature.IsAvailable)
                 {
-                    double temperature = cpuTemperature.Temperature.Celsius;
+                    double temperature = cpuTemperature.Temperature.DegreesCelsius;
                     if (!double.IsNaN(temperature))
                     {
                         Console.WriteLine($"CPU Temperature: {temperature} C");
