@@ -27,14 +27,14 @@ namespace Iot.Device.DHTxx.Samples
                     var tempValue = dht.Temperature;
                     var humValue = dht.Humidity;
 
-                    Console.WriteLine($"Temperature: {tempValue.Celsius:0.#}\u00B0C");
+                    Console.WriteLine($"Temperature: {tempValue.DegreesCelsius:0.#}\u00B0C");
                     Console.WriteLine($"Relative humidity: {humValue:0.#}%");
 
                     // WeatherHelper supports more calculations, such as saturated vapor pressure, actual vapor pressure and absolute humidity.
                     Console.WriteLine(
-                        $"Heat index: {WeatherHelper.CalculateHeatIndex(tempValue, humValue).Celsius:0.#}\u00B0C");
+                        $"Heat index: {WeatherHelper.CalculateHeatIndex(tempValue, humValue).DegreesCelsius:0.#}\u00B0C");
                     Console.WriteLine(
-                        $"Dew point: {WeatherHelper.CalculateDewPoint(tempValue, humValue).Celsius:0.#}\u00B0C");
+                        $"Dew point: {WeatherHelper.CalculateDewPoint(tempValue, humValue).DegreesCelsius:0.#}\u00B0C");
 
                     Thread.Sleep(2000);
                 }
