@@ -4,7 +4,7 @@
 
 using System;
 using System.Device.I2c;
-using Iot.Units;
+using UnitsNet;
 
 namespace Iot.Device.Lm75
 {
@@ -25,7 +25,7 @@ namespace Iot.Device.Lm75
         /// <summary>
         /// LM75 Temperature
         /// </summary>
-        public Temperature Temperature { get => Temperature.FromCelsius(GetTemperature()); }
+        public Temperature Temperature { get => Temperature.FromDegreesCelsius(GetTemperature()); }
 
         private bool _disable;
 
