@@ -12,7 +12,7 @@ using System.Numerics;
 using System.Text;
 using System.Threading;
 using Iot.Device.Magnetometer;
-using Iot.Units;
+using UnitsNet;
 
 namespace Iot.Device.Imu
 {
@@ -340,7 +340,7 @@ namespace Iot.Device.Imu
             };
             ReadBytes(Register.TEMP_OUT_H, rawData);
             // formula from the documentation
-            return Temperature.FromCelsius((BinaryPrimitives.ReadInt16BigEndian(rawData) - 21) / 333.87 + 21);
+            return Temperature.FromDegreesCelsius((BinaryPrimitives.ReadInt16BigEndian(rawData) - 21) / 333.87 + 21);
         }
 
         #endregion
